bot_service/pyproject.toml
[project]
name = "bot-service"
version = "1.0.0"
description = "телеграм бот"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.10.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "fakeredis[lua]>=2.21.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "fakeredis[lua]>=2.21.0",
    "respx>=0.21.0",
]
